Move x509_certificate_info code to module_utils.
